+++ src/views/admin-component-settings/site.php
@@ -34,8 +34,29 @@
 
 
 <?php $form = \skeeks\cms\modules\admin\widgets\form\ActiveFormUseTab::begin(); ?>
-<?= $component->renderConfigForm($form); ?>
+
+<?= $form->errorSummary(\yii\helpers\ArrayHelper::merge(
+        [$component], $component->getConfigFormModels()
+)); ?>
+
+<? if ($fields = $component->getConfigFormFields()) : ?>
+    <? echo (new \skeeks\yii2\form\Builder([
+        'models'     => $component->getConfigFormModels(),
+        'model'      => $component,
+        'activeForm' => $form,
+        'fields'     => $fields,
+    ]))->render(); ?>
+<? elseif ($formContent = $component->renderConfigForm($form)) : ?>
+    <?= $formContent; ?>
+<? else : ?>
+    Нет редактируемых настроек для данного компонента
+<? endif; ?>
+
 <?= $form->buttonsStandart($component); ?>
+<?= $form->errorSummary(\yii\helpers\ArrayHelper::merge(
+        [$component], $component->getConfigFormModels()
+)); ?>
+
 <?php \skeeks\cms\modules\admin\widgets\form\ActiveFormUseTab::end(); ?>